requirements.txt
aiohttp==1.0.5
beautifulsoup4==4.5.1
chardet==2.3.0
discord.py==0.13.0
lxml==3.6.4
peewee==2.8.5
websockets==3.2
pytz==2016.7
